Added shift option to processor instructions
Added the option to specify a shift in the operations performed by the
processor, so that the inputs used for the pixel at (x, y) come from
(x+k, y) if x+k < width or from (width-1, y) otherwise.
Added a testbench to test edge detection using shifting, absolute
difference and threshold.
Fixed an implicit latch of error_code by adding it to the synchronous
block.